refactor(BridgeMaker): 디렉토리 원래대로 수정
