+++ src/main/java/bridge/controller/BridgeController.java
@@ -2,7 +2,7 @@ package bridge.controller;
 
 import bridge.BridgeRandomNumberGenerator;
 import bridge.domain.BridgeGame;
-import bridge.domain.BridgeMaker;
+import bridge.BridgeMaker;
 import bridge.ui.InputView;
 import bridge.ui.OutputView;
 import java.util.List;
